devise-eye/requirements.txt
psutil>=5.9.0
dnspython>=2.4.0
APScheduler>=3.10.0
httpx>=0.25.0
python-dotenv>=1.0.0
PyInstaller>=6.0.0
keyring>=24.0.0
google-auth>=2.23.0
pywin32>=306
pyperclip>=1.8.2
